Add BranchManagement to sidebar/BrandManagementController
