+++ FMB-CIS/FMB-CIS/Controllers/OfficeManagementController.cs
@@ -682,257 +682,5 @@ namespace FMB_CIS.Controllers
             return Json(regionLists);
         }
 
-
-        /// <summary>
-        /// Brands
-        /// </summary>
-        /// <returns></returns>
-        public IActionResult BrandList()
-        {
-            int uid = Convert.ToInt32(((ClaimsIdentity)User.Identity).FindFirst("userID").Value);
-            int usrRoleID = _context.tbl_user.Where(u => u.id == uid).Select(u => u.tbl_user_types_id).SingleOrDefault();
-            if (usrRoleID == 14) // Super Admin
-            {
-                string host = $"{Request.Scheme}://{Request.Host}{Request.PathBase}/";
-                ViewData["BaseUrl"] = host;
-
-                return View();
-            }
-            else
-            {
-                return RedirectToAction("Index", "Dashboard");
-            }
-        }
-
-        public IActionResult BrandListPartialView()
-        {
-            int uid = Convert.ToInt32(((ClaimsIdentity)User.Identity).FindFirst("userID").Value);
-            int usrRoleID = _context.tbl_user.Where(u => u.id == uid).Select(u => u.tbl_user_types_id).SingleOrDefault();
-            if (usrRoleID == 14) // Super Admin
-            {
-                BrandListViewModel model = new BrandListViewModel();
-                //Get the list of users
-                var entities = _context.tbl_brands.Where(e => e.is_active == true).ToList();
-                model.brands = entities.Adapt<List<Brand>>();
-
-                string host = $"{Request.Scheme}://{Request.Host}{Request.PathBase}/";
-                ViewData["BaseUrl"] = host;
-
-                return PartialView("~/Views/OfficeManagement/Brand/Partial/BrandListPartial.cshtml", model);
-            }
-            else
-            {
-                return RedirectToAction("Index", "Dashboard");
-            }
-        }
-
-
-        public IActionResult BrandCreateModal()
-        {
-            int uid = Convert.ToInt32(((ClaimsIdentity)User.Identity).FindFirst("userID").Value);
-            int usrRoleID = _context.tbl_user.Where(u => u.id == uid).Select(u => u.tbl_user_types_id).SingleOrDefault();
-            if (usrRoleID == 14) // Super Admin
-            {
-                BrandListViewModel model = new BrandListViewModel();
-                //Get the list of users
-                //var entities = _context.tbl_office_type.ToList();
-                //model.officeTypes = entities.Adapt<List<OfficeType>>();
-
-                string host = $"{Request.Scheme}://{Request.Host}{Request.PathBase}/";
-                ViewData["BaseUrl"] = host;
-
-                return PartialView("~/Views/OfficeManagement/Brand/Modal/BrandCreateModal.cshtml", model);
-            }
-            else
-            {
-                return RedirectToAction("Index", "Dashboard");
-            }
-        }
-        public IActionResult BrandUpdateModal(int id)
-        {
-            int uid = Convert.ToInt32(((ClaimsIdentity)User.Identity).FindFirst("userID").Value);
-            int usrRoleID = _context.tbl_user.Where(u => u.id == uid).Select(u => u.tbl_user_types_id).SingleOrDefault();
-            if (usrRoleID == 14) // Super Admin
-            {
-                Brand model = new Brand();
-                //Get the list of users
-                var entity = _context.tbl_brands.FirstOrDefault(o => o.id == id);
-                model = entity.Adapt<Brand>();
-
-                string host = $"{Request.Scheme}://{Request.Host}{Request.PathBase}/";
-                ViewData["BaseUrl"] = host;
-
-                return PartialView("~/Views/OfficeManagement/Brand/Modal/BrandUpdateModal.cshtml", model);
-            }
-            else
-            {
-                return RedirectToAction("Index", "Dashboard");
-            }
-        }
-        public IActionResult BrandDeleteModal(int id)
-        {
-            int uid = Convert.ToInt32(((ClaimsIdentity)User.Identity).FindFirst("userID").Value);
-            int usrRoleID = _context.tbl_user.Where(u => u.id == uid).Select(u => u.tbl_user_types_id).SingleOrDefault();
-            if (usrRoleID == 14) // Super Admin
-            {
-                Brand model = new Brand();
-                //Get the list of users
-                var brand = _context.tbl_brands.FirstOrDefault(o => o.id == id);
-                model = brand.Adapt<Brand>();
-
-                string host = $"{Request.Scheme}://{Request.Host}{Request.PathBase}/";
-                ViewData["BaseUrl"] = host;
-
-                return PartialView("~/Views/OfficeManagement/OfficeType/Modal/BrandDeleteModal.cshtml", model);
-            }
-            else
-            {
-                return RedirectToAction("Index", "Dashboard");
-            }
-        }
-
-        [HttpPost]
-        public IActionResult BrandCreate(BrandCreateViewModel model)
-        {
-            int uid = Convert.ToInt32(((ClaimsIdentity)User.Identity).FindFirst("userID").Value);
-            int usrRoleID = _context.tbl_user.Where(u => u.id == uid).Select(u => u.tbl_user_types_id).SingleOrDefault();
-            // Note: TO DO Soon move Authorization to middleware / Annotations / User Access instead of Roles
-            if (usrRoleID == 14) // Super Admin
-            {
-                // Uses fluent validation
-                if (!ModelState.IsValid)
-                {
-                    return StatusCode(StatusCodes.Status400BadRequest, ModelState);
-                }
-                else
-                {
-                    // Save
-                    // Note: TO DO Soon refactor code, let's move the Data Access to a new layer
-
-                    // Action: 1 to 5
-                    // 1. create entity
-                    var entity = new tbl_brand();
-                    entity.name = model.name;
-                    entity.description = model.description;
-                    entity.is_active = true;
-                    entity.createdBy = uid;
-                    entity.modifiedBy = uid;
-                    entity.date_created = DateTime.Now;
-                    entity.date_modified = DateTime.Now;
-
-                    // 2. add to context
-                    _context.tbl_brands.Add(entity);
-
-                    // 3. TO DO: Add logging / historical data
-
-                    // 4. Save changes
-                    _context.SaveChanges();
-
-                    // 5. Return result
-                    return StatusCode(StatusCodes.Status201Created, ModelState);
-                }
-            }
-            else
-            {
-                return RedirectToAction("Index", "AccountManagement");
-            }
-        }
-        [HttpPut("OfficeManagement/BrandUpdate/{id:int}")]
-        public IActionResult BrandUpdate(BrandUpdateViewModel model)
-        {
-            int uid = Convert.ToInt32(((ClaimsIdentity)User.Identity).FindFirst("userID").Value);
-            int usrRoleID = _context.tbl_user.Where(u => u.id == uid).Select(u => u.tbl_user_types_id).SingleOrDefault();
-            // Note: TO DO Soon move Authorization to middleware / Annotations / User Access instead of Roles
-            if (usrRoleID == 14) // Super Admin
-            {
-                // Uses fluent validation
-                if (!ModelState.IsValid)
-                {
-                    return StatusCode(StatusCodes.Status400BadRequest, ModelState);
-                }
-                else
-                {
-                    // Save
-                    // Note: TO DO Soon refactor code, let's move the Data Access to a new layer
-
-                    // Action: 1 to 5
-                    // 1. get, update entity
-                    var entity = _context.tbl_brands.Where(m => m.id == model.id).FirstOrDefault();
-
-                    // NOTE: TO DO, is there a better location for this
-                    if (entity == null)
-                    {
-                        return StatusCode(StatusCodes.Status404NotFound, ModelState);
-                    }
-                    entity.name = model.name;
-                    entity.description = model.description;
-                    entity.modifiedBy = uid;
-                    entity.date_modified = DateTime.Now;
-
-                    // 2. update to context
-                    _context.Update(entity);
-
-                    // 3. TO DO: Add logging / historical data
-
-                    // 4. Save changes
-                    _context.SaveChanges();
-
-                    // 5. Return result
-                    return StatusCode(StatusCodes.Status201Created, ModelState);
-                }
-            }
-            else
-            {
-                return RedirectToAction("Index", "AccountManagement");
-            }
-        }
-        [HttpDelete("OfficeManagement/BrandDelete/{id:int}")]
-        public IActionResult BrandDelete(BrandDeleteViewModel model)
-        {
-            int uid = Convert.ToInt32(((ClaimsIdentity)User.Identity).FindFirst("userID").Value);
-            int usrRoleID = _context.tbl_user.Where(u => u.id == uid).Select(u => u.tbl_user_types_id).SingleOrDefault();
-            // Note: TO DO Soon move Authorization to middleware / Annotations / User Access instead of Roles
-            if (usrRoleID == 14) // Super Admin
-            {
-                // Uses fluent validation
-                if (!ModelState.IsValid)
-                {
-                    return StatusCode(StatusCodes.Status400BadRequest, ModelState);
-                }
-                else
-                {
-                    // Save
-                    // Note: TO DO Soon refactor code, let's move the Data Access to a new layer
-
-                    // Action: 1 to 5
-                    // 1. get, update entity
-                    var entity = _context.tbl_brands.Where(m => m.id == model.id).FirstOrDefault();
-
-                    // NOTE: TO DO, is there a better location for this
-                    if (entity == null)
-                    {
-                        return StatusCode(StatusCodes.Status404NotFound, ModelState);
-                    }
-                    entity.is_active = false;
-                    entity.modifiedBy = uid;
-                    entity.date_modified = DateTime.Now;
-
-                    // 2. update to context
-                    _context.Update(entity);
-
-                    // 3. TO DO: Add logging / historical data
-
-                    // 4. Save changes
-                    _context.SaveChanges();
-
-                    // 5. Return result
-                    return StatusCode(StatusCodes.Status201Created, ModelState);
-                }
-            }
-            else
-            {
-                return RedirectToAction("Index", "AccountManagement");
-            }
-        }
     }
 }
