+++ FMB-CIS/FMB-CIS/Views/BrandManagement/Brand/Modal/BrandDeleteModal.cshtml
@@ -1,0 +1,48 @@
+﻿@using System.Security.Claims;
+@using System.Globalization;
+@model FMB_CIS.Models.Brand
+@{
+    ViewBag.Title = "Brand Management";
+}
+<div class="modal modal-sheet  bg-body-secondary p-4 py-md-5" tabindex="-1" role="dialog" id="deleteBrandModal" aria-hidden="true">
+    <div class="modal-dialog" role="document">
+        <div class="modal-content rounded-3 shadow">
+            <div class="modal-body p-4 text-center">
+                <h5 class="mb-0">Are you sure with deleting this item?</h5>
+                <p class="mb-0">This will render the connected properties be inactive as well.</p>
+            </div>
+            <div class="modal-footer flex-nowrap p-0">
+                <input type="hidden" class="form-control rounded-3" id="brandId" asp-for="@Model.id">
+                <button id="btnDelete" type="button" class="btn btn-danger btn-lg fs-6 text-decoration-none col-6 py-3 m-0 rounded-0 border-end"><strong>Yes, proceed deleting</strong></button>
+                <button id="cancelBtn" type="button" class="btn btn-lg btn-link fs-6 text-decoration-none col-6 py-3 m-0 rounded-0" data-bs-dismiss="modal">No, close</button>
+            </div>
+        </div>
+    </div>
+</div>
+
+<script type="text/javascript">
+    $("#cancelBtn").click(function () {
+        closeModal();
+    });
+
+    $("#btnDelete").click(function () {
+        console.log("btnDelete");
+        let id = document.getElementById("brandId").value;
+        let parameters = { id: id};
+        $.ajax({
+            url: '/OfficeManagement/BrandDelete/' + id,
+            type: 'DELETE',
+            data: parameters,
+            success: function (data) {
+                // NOTE: TO DO better way of calling callback function from parent
+                $('#brandPartialContainer').load('/OfficeManagement/BrandListPartialView');
+            }
+        });
+    });
+
+    function closeModal() {
+        $("#deleteBrandModal").modal("hide");
+        let elem = document.querySelector('#deleteBrandModal');
+        elem.parentNode.removeChild(elem);
+    }
+</script>
