+++ FMB-CIS/FMB-CIS/Views/BrandManagement/Brand/Modal/BrandUpdateModal.cshtml
@@ -1,0 +1,64 @@
+﻿@using System.Security.Claims;
+@using System.Globalization;
+@model FMB_CIS.Models.Brand
+@{
+    ViewBag.Title = "Brand Management";
+}
+
+<div class="modal modal-sheet fade  bg-body-secondary p-4 py-md-5" id="updateBrandModal" aria-hidden="true">
+    <div class="modal-dialog" role="document">
+        <div class="modal-content rounded-4 shadow">
+            <div class="modal-header p-5 pb-4 border-bottom-0">
+                <h1 class="fw-bold mb-0 fs-2">Update Brand</h1>
+                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+            </div>
+
+            <div class="modal-body p-5 pt-0">
+                <form class="">
+                    <input type="hidden" class="form-control rounded-3" id="brandId" asp-for="@Model.id">
+                    <div class="form-floating mb-3">
+                        <input type="text" class="form-control rounded-3" id="brandName" placeholder="" asp-for="@Model.name">
+                        <label for="brandName">Brand Name</label>
+                    </div>
+                    <div class="form-floating mb-3">
+                        <input type="text" class="form-control rounded-3" id="brandDescription" placeholder="" asp-for="@Model.description">
+                        <label for="brandDescription">Brand Description</label>
+                    </div>
+                    <button id="btnUpdate" class="w-100 mb-2 btn btn-lg rounded-3 btn-primary" type="button">Update</button>
+                    <small class="text-body-secondary">By clicking Update, the brand will be updated on the system.</small>
+                </form>
+            </div>
+        </div>
+    </div>
+</div>
+<script type="text/javascript">
+    $("#cancelBtn").click(function () {
+        closeModal();
+    });
+
+    $("#btnUpdate").click(function () {
+        console.log("btnUpdate");
+        let id = document.getElementById("brandId").value;
+        let parameters = {
+            id: id,
+            name: document.getElementById("brandName").value,
+            description: document.getElementById("brandDescription").value
+        };
+        $.ajax({
+            url: '/OfficeManagement/BrandUpdate/' + id,
+            type: 'PUT',
+            data: parameters,
+            success: function (data) {
+                // NOTE: TO DO better way of calling callback function from parent
+                $('#brandListPartialContainer').load('/OfficeManagement/BrandListPartialView');
+                closeModal();
+            }
+        });
+    });
+
+    function closeModal() {
+        $("#updateBrandModal").modal("hide");
+        let elem = document.querySelector('#updateBrandModal');
+        elem.parentNode.removeChild(elem);
+    }
+</script>
