+++ FMB-CIS/FMB-CIS/Views/BrandManagement/BrandList.cshtml
@@ -1,0 +1,45 @@
+﻿@using System.Security.Claims;
+@using System.Globalization;
+@{
+    Layout = "~/Views/Shared/_MainPageLayout.cshtml";
+    ViewBag.Title = "Brand Management";
+}
+
+<div class="c-card-container c-card-container-box">
+    <div id="card">
+        <div class="c-info">
+            <div class="card">
+                <div class="card-body">
+                    <h5 class="card-title">
+                        <a href="@Url.Action("Index", "Home")" class="me-auto" title="back"><i data-feather="arrow-left" class="text-secondary backBtn"></i></a>
+                        Brands
+                    </h5>
+                    <p class="card-text">Used to represent different types of Chainsaw Brands.</p>
+                    <button id="btnCreateBrand" onclick="OpenCreateModal()" class="btn btn-primary">Create New Brand</button>
+                </div>
+            </div>
+        </div>
+        <div class="c-content" id="brandListPartialContainer">
+            <div class="d-flex justify-content-center">
+                <div class="spinner-border m-5" role="status">
+                    <span class="visually-hidden">Loading...</span>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+
+<script src="/js/jquery-3.7.1.min.js"></script>
+
+<script type="text/javascript">
+    $("#btnCreateBrand").click(function () {
+    });
+
+    function OpenCreateModal() {
+        $('#brandCreateModalContainer').load('/BrandManagement/BrandCreateModal', function () {
+            $("#createBrandModal").modal("show");
+        });
+    }
+
+    $('#brandListPartialContainer').load('/BrandManagement/BrandListPartialView');
+</script>
