+++ FMB-CIS/FMB-CIS/wwwroot/css/main-page-layout.css
@@ -82,14 +82,30 @@
         color: #14984a;
     }
 
-#chainsawImporterDropdown, #chainsawSellerDropdown, #chainsawOwnerDropdown, #authorityToLeaseDropdown, #accountManagementDropdown, #officeManagementDropdown, #accessRightsManagementDropdown, #workflowManagementDropdown {
+#chainsawImporterDropdown, 
+#chainsawSellerDropdown, 
+#chainsawOwnerDropdown, 
+#authorityToLeaseDropdown, 
+#accountManagementDropdown, 
+#officeManagementDropdown, 
+#brandManagementDropdown, 
+#accessRightsManagementDropdown, 
+#workflowManagementDropdown {
     max-height: 0;
     overflow: hidden;
     opacity: 0;
     transition: all 0.15s ease-in-out;
 }
 
-    #chainsawImporterDropdown.active, #chainsawSellerDropdown.active, #chainsawOwnerDropdown.active, #authorityToLeaseDropdown.active, #accountManagementDropdown.active, #officeManagementDropdown.active, #accessRightsManagementDropdown.active, #workflowManagementDropdown.active {
+    #chainsawImporterDropdown.active, 
+    #chainsawSellerDropdown.active, 
+    #chainsawOwnerDropdown.active, 
+    #authorityToLeaseDropdown.active, 
+    #accountManagementDropdown.active, 
+    #officeManagementDropdown.active, 
+    #brandManagementDropdown.active, 
+    #accessRightsManagementDropdown.active, 
+    #workflowManagementDropdown.active {
         max-height: 100%;
         opacity: 1;
         margin-top: 15px;
